+++ CHANGELOG.md
@@ -3,7 +3,17 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-## [0.1.0] - 2020-04-26
+## [0.1.4] - 2021-06-05
+### Added
+- A separate evaluation for binary classification problem
+- Functions documentation
+- Automatic generation for `setup.py` file
+
+### Fixed
+- Fix `_formula` function error for matrix multiplication
+
+
+## [0.1.0] - 2021-04-26
 ### Added
 - Compute confusion matrix using `sklearn`
 - Transform multi-label classification to binary classification